requirements.txt
# Core dependencies
selenium==4.27.1
webdriver-manager==4.0.2

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Web framework
flask==3.1.0
flask-cors==5.0.0

# Data processing
pandas==2.2.3
numpy==2.2.1
beautifulsoup4==4.12.3
strawberry-graphql==0.208.0
graphql-core==3.2.5
lxml==5.3.0

# API & HTTP
requests==2.32.3
urllib3==2.2.3

# Configuration
pyyaml==6.0.2
python-dotenv==1.0.1

# Validation
validators==0.34.0
phonenumbers==8.13.50

# Export formats
openpyxl==3.1.5
xlsxwriter==3.2.0

# LLM Integration
ollama==0.4.4

# Visualization
plotly==5.24.1
matplotlib==3.10.0

# Logging & monitoring
colorlog==6.9.0

# Collaborative intelligence
pinecone-client==3.0.0
langchain==0.1.0
web3==6.15.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0

# Utilities
tqdm==4.67.1
schedule==1.2.2
